requirements.txt
flask==2.3.3
flask-mail==0.9.1
pymongo==4.5.0
python-dotenv==1.0.0
pandas==2.0.3
scikit-learn==1.3.1
joblib==1.3.2
nltk==3.8.1
google-generativeai==0.3.0
werkzeug==2.3.7
requests==2.31.0
